# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Thanks to Cerebras API",
    "version": "1.0.0",
    "description": "OpenAPI schema for the OpenAI-compatible proxy and same-origin admin API. Every response includes x-request-id; clients may send X-Request-Id to propagate a correlation ID."
  },
  "servers": [{ "url": "/" }],
...
  "paths": {
...
    "/api/proxy-keys": {
      "get": {
        "tags": ["Admin proxy keys"],
        "operationId": "listProxyKeys",
        "summary": "List proxy key metadata",
        "security": [{ "AdminToken": [] }],
        "responses": {
          "200": {
            "description": "Proxy key metadata list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProxyKeyList" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Problem" },
          "500": { "$ref": "#/components/responses/Problem" }
        }
      },
      "post": {
        "tags": ["Admin proxy keys"],
        "operationId": "createProxyKey",
        "summary": "Create a proxy key; plaintext is returned once",
        "security": [{ "AdminToken": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProxyKeyRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Proxy key created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProxyKeyResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Problem" },
          "401": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/api/proxy-keys/migrate": {
      "post": {
        "tags": ["Admin proxy keys"],
        "operationId": "migrateProxyKeys",
        "summary": "Migrate proxy keys to hashed storage",
        "security": [{ "AdminToken": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/Migrated" },
          "401": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/api/proxy-keys/{id}": {
      "delete": {
        "tags": ["Admin proxy keys"],
        "operationId": "deleteProxyKey",
        "summary": "Delete a proxy key",
        "security": [{ "AdminToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "400": { "$ref": "#/components/responses/Problem" },
          "401": { "$ref": "#/components/responses/Problem" },
          "404": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/api/proxy-keys/{id}/export": {
      "get": {
        "tags": ["Admin proxy keys"],
        "operationId": "exportProxyKey",
        "summary": "Reject proxy key export after creation",
        "security": [{ "AdminToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": { "403": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/api/models/catalog": {
      "get": {
        "tags": ["Admin models"],
        "operationId": "getModelCatalog",
        "summary": "Get the cached Cerebras public model catalog",
        "security": [{ "AdminToken": [] }],
        "responses": {
          "200": {
            "description": "Model catalog",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ModelCatalog" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Problem" },
          "502": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/api/models/catalog/refresh": {
      "post": {
        "tags": ["Admin models"],
        "operationId": "refreshModelCatalog",
        "summary": "Refresh the Cerebras public model catalog",
        "security": [{ "AdminToken": [] }],
        "responses": {
          "200": {
            "description": "Refreshed or stale fallback model catalog",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ModelCatalog" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Problem" },
          "502": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/api/models": {
      "get": {
        "tags": ["Admin models"],
        "operationId": "getModelPool",
        "summary": "Get the active model pool",
        "security": [{ "AdminToken": [] }],
        "responses": {
          "200": {
            "description": "Active model pool",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ModelPool" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Problem" }
        }
      },
      "put": {
        "tags": ["Admin models"],
        "operationId": "updateModelPool",
        "summary": "Replace the active model pool",
        "security": [{ "AdminToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateModelPoolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated model pool",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateModelPoolResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Problem" },
          "401": { "$ref": "#/components/responses/Problem" },
          "500": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/api/models/{name}/test": {
      "post": {
        "tags": ["Admin models"],
        "operationId": "testModel",
        "summary": "Test whether a model is usable with an active API key",
        "security": [{ "AdminToken": [] }],
        "parameters": [{
          "name": "name",
          "in": "path",
          "required": true,
          "description": "URL-encoded model name",
          "schema": { "type": "string" }
        }],
        "responses": {
          "200": {
            "description": "Model test result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TestResult" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Problem" },
          "401": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token"
      },
      "SetupToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Setup-Token"
      },
      "ProxyBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Proxy key supplied as Authorization: Bearer <proxy-key>. May be bypassed when proxyPublicAccess is enabled."
      }
    },
    "headers": {
      "XRequestId": {
        "description": "Correlation ID echoed from X-Request-Id or generated per request",
        "schema": { "type": "string" }
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
    },
...
    "responses": {
      "Problem": {
        "description": "Problem Details error",
        "headers": {
          "x-request-id": { "$ref": "#/components/headers/XRequestId" }
        },
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
...
      "Success": {
        "description": "Successful operation",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Success" }
          }
        }
      },
      "Migrated": {
        "description": "Migration completed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Migrated" }
          }
        }
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "required": ["type", "title", "status", "detail"],
        "properties": {
          "type": { "type": "string" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "instance": { "type": "string" }
        },
        "additionalProperties": false
      },
...
      "Success": {
        "type": "object",
        "required": ["success"],
        "properties": { "success": { "type": "boolean", "const": true } },
        "additionalProperties": true
      },
      "Migrated": {
        "type": "object",
        "required": ["success", "migrated"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "migrated": { "type": "integer", "minimum": 0 }
        },
        "additionalProperties": false
      },
...
      "ProxyKeyList": {
        "type": "object",
        "required": ["keys", "maxKeys", "authEnabled", "proxyPublicAccess"],
        "properties": {
          "keys": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProxyKeyMetadata" }
          },
          "maxKeys": { "type": "integer" },
          "authEnabled": { "type": "boolean", "const": true },
          "proxyPublicAccess": { "type": "boolean" }
        },
        "additionalProperties": false
      },
      "ProxyKeyMetadata": {
        "type": "object",
        "required": ["id", "name", "useCount", "createdAt"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "useCount": { "type": "integer" },
          "lastUsed": { "type": "integer" },
          "createdAt": { "type": "integer" }
        },
        "additionalProperties": false
      },
      "CreateProxyKeyRequest": {
        "type": "object",
        "properties": { "name": { "type": "string" } },
        "additionalProperties": false
      },
      "CreateProxyKeyResponse": {
        "type": "object",
        "required": ["success", "id", "key"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "id": { "type": "string" },
          "key": { "type": "string" }
        },
        "additionalProperties": false
      },
      "ModelCatalog": {
        "type": "object",
        "required": ["source", "fetchedAt", "ttlMs", "stale", "models"],
        "properties": {
          "source": { "type": "string", "const": "cerebras-public" },
          "fetchedAt": { "type": "integer" },
          "ttlMs": { "type": "integer" },
          "stale": { "type": "boolean" },
          "lastError": { "type": "string" },
          "models": { "type": "array", "items": { "type": "string" } }
        },
        "additionalProperties": false
      },
      "ModelPool": {
        "type": "object",
        "required": ["models"],
        "properties": {
          "models": { "type": "array", "items": { "type": "string" } }
        },
        "additionalProperties": false
      },
      "UpdateModelPoolRequest": {
        "type": "object",
        "required": ["models"],
        "properties": {
          "models": {
            "type": "array",
            "minItems": 1,
            "items": { "type": "string" }
          }
        },
        "additionalProperties": false
      },
      "UpdateModelPoolResponse": {
        "type": "object",
        "required": ["success", "models"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "models": { "type": "array", "items": { "type": "string" } }
        },
        "additionalProperties": false
      },
      "TestResult": {
        "type": "object",
        "required": ["success", "status"],
        "properties": {
          "success": { "type": "boolean" },
          "status": { "type": "string" },
          "error": { "type": "string" }
        },
        "additionalProperties": false
      },
...
    }
  }
}